refactor: moved artist components into home folder
