+++ frontend/src/pages/Home/Artists/Artists.css
@@ -1,7 +1,3 @@
-.artists-body {
-    padding: 20px;
-}
-
 .artist {
     margin-bottom: 20px;
     display: flex;
